# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: NyelvSzó API
  description: |
    Angol-magyar Nyelvészeti Szakszótár API dokumentáció.
    
    Ez az API lehetővé teszi a nyelvészeti szakkifejezések keresését, létrehozását, módosítását és törlését.
    Három felhasználói szerepkör létezik:
    - **Felhasználó (1)**: Csak megtekintés
    - **Szerkesztő (2)**: Szócikkek kezelése
    - **Adminisztrátor (3)**: Teljes hozzáférés
    
    ## Autentikáció
    Az API JWT (JSON Web Token) alapú autentikációt használ. A token-t a `/login` végponton lehet megszerezni.
    
    ## Rate Limiting
    - Általános végpontok: 100 kérés / 15 perc / IP
    - Login végpont: 5 kérés / 15 perc / IP
    
  version: 2.2.0
  contact:
    name: Dr. Ádám Porkoláb
    email: adam@porkolab.digital
    url: https://nyelvszo.eu
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.nyelvszo.eu
    description: Production server
paths:
...
  /entries:
    get:
      tags:
        - Entries
      summary: Get all entries
      description: Retrieve all dictionary entries with optional filtering
      operationId: getAllEntries
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          description: Search term for filtering entries
          schema:
            type: string
        - name: fieldOfExpertise
          in: query
          description: Filter by field of expertise
          schema:
            type: string
        - name: wordType
          in: query
          description: Filter by word type
          schema:
            type: string
      responses:
        '200':
          description: List of entries retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Entry'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Entries
      summary: Create new entry
      description: Create a new dictionary entry (requires editor role or higher)
      operationId: createEntry
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntryInput'
      responses:
        '201':
          description: Entry created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /entries/{id}:
    get:
      tags:
        - Entries
      summary: Get entry by ID
      description: Retrieve a specific dictionary entry by its ID
      operationId: getEntryById
      parameters:
        - name: id
          in: path
          required: true
          description: Entry ID
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: Entry retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
        '404':
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Entries
      summary: Update entry
      description: Update an existing dictionary entry (requires editor role or higher)
      operationId: updateEntry
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Entry ID
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntryInput'
      responses:
        '200':
          description: Entry updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Entries
      summary: Partially update entry
      description: Partially update an existing dictionary entry (requires editor role or higher)
      operationId: patchEntry
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Entry ID
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntryUpdateInput'
      responses:
        '200':
          description: Entry updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Entries
      summary: Delete entry
      description: Delete a dictionary entry (requires editor role or higher)
      operationId: deleteEntry
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Entry ID
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: Entry deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Entry deleted successfully"
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT authentication. Add 'Bearer ' prefix to your JWT token.
        Example: `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
    Entry:
      type: object
      properties:
        _id:
          type: string
          format: objectId
          description: Unique identifier
          example: "507f1f77bcf86cd799439011"
        hungarian:
          type: string
          description: Hungarian term
          example: "számítógép"
          maxLength: 500
        english:
          type: string
          description: English term
          example: "computer"
          maxLength: 500
        fieldOfExpertise:
          type: string
          description: Field of expertise
          example: "informatika"
          maxLength: 200
        wordType:
          type: string
          description: Type of word (noun, verb, etc.)
          example: "főnév"
          maxLength: 100
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - _id
        - hungarian
        - english
        - fieldOfExpertise
    EntryInput:
      type: object
      properties:
        hungarian:
          type: string
          description: Hungarian term
          example: "számítógép"
          minLength: 1
          maxLength: 500
        english:
          type: string
          description: English term
          example: "computer"
          minLength: 1
          maxLength: 500
        fieldOfExpertise:
          type: string
          description: Field of expertise
          example: "informatika"
          minLength: 1
          maxLength: 200
        wordType:
          type: string
          description: Type of word (noun, verb, etc.)
          example: "főnév"
          minLength: 1
          maxLength: 100
      required:
        - hungarian
        - english
        - fieldOfExpertise
    EntryUpdateInput:
      type: object
      properties:
        hungarian:
          type: string
          description: Hungarian term
          example: "számítógép"
          minLength: 1
          maxLength: 500
        english:
          type: string
          description: English term
          example: "computer"
          minLength: 1
          maxLength: 500
        fieldOfExpertise:
          type: string
          description: Field of expertise
          example: "informatika"
          minLength: 1
          maxLength: 200
        wordType:
          type: string
          description: Type of word (noun, verb, etc.)
          example: "főnév"
          minLength: 1
          maxLength: 100
...
    Pagination:
      type: object
      properties:
        currentPage:
          type: integer
          description: Current page number
          example: 1
        totalPages:
          type: integer
          description: Total number of pages
          example: 10
        totalItems:
          type: integer
          description: Total number of items
          example: 200
        itemsPerPage:
          type: integer
          description: Number of items per page
          example: 20
        hasNextPage:
          type: boolean
          description: Whether there is a next page
          example: true
        hasPrevPage:
          type: boolean
          description: Whether there is a previous page
          example: false
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid credentials"
        message:
          type: string
          description: Detailed error message
          example: "The provided email or password is incorrect"
        statusCode:
          type: integer
          description: HTTP status code
          example: 401
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        path:
          type: string
          description: Request path where error occurred
          example: "/login"
      required:
        - error
        - statusCode
    ValidationError:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Validation Error"
        message:
          type: string
          description: Detailed error message
          example: "Input validation failed"
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
        details:
          type: array
          description: Validation error details
          items:
            type: object
            properties:
              field:
                type: string
                description: Field that failed validation
                example: "email"
              message:
                type: string
                description: Validation error message
                example: "Please provide a valid email address"
              value:
                description: The invalid value that was provided
                example: "invalid-email"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        path:
          type: string
          description: Request path where error occurred
          example: "/users"
      required:
        - error
        - statusCode
        - details
...